+++ src/app/collective-rewards/hooks/index.ts
@@ -1,1 +1,1 @@
-export * from './useGetGaugesFunction'
+export * from './useGaugesGetFunction'
