+++ src/app/collective-rewards/hooks/useGaugesGetFunction.ts
